streams/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
SQLAlchemy==1.4.37
